+++ .eslintrc.js
@@ -24,6 +24,6 @@ module.exports = {
   rules: {
     'prefer-named-capture-group': 0,
     'react/no-did-mount-set-state': 1,
-    'react/prop-types': 1,
+    'react/prop-types': 0,
   },
 }
