+++ src/pages/ProductList/index.js
@@ -6,27 +6,18 @@ import Loader from '../../components/Loader'
 import { H1 } from '../../components/Typography'
 
 import { getProducts } from '../../api/get-products'
-import { addProduct } from '../../store/cartItems/actions'
+import { addProduct } from '../../store/cart/actions'
 import { loadProducts } from '../../store/products/actions'
 import Product from './Product'
 import { ProductsWrap } from './styled'
 
 class Products extends Component {
-  state = {
-    isLoading: true,
-  }
-
   async componentDidMount() {
     const products = await getProducts()
     this.props.loadProducts(products)
-
-    this.setState({
-      isLoading: false,
-    })
   }
 
-  handleAddToCart = (productId, evt) => {
-    evt.preventDefault()
+  handleAddToCart = (productId) => {
     this.props.addProduct(productId)
   }
 
@@ -34,7 +25,7 @@ class Products extends Component {
     return (
       <Layout>
         <H1 textAlign="center">E-Commerce app</H1>
-        {this.state.isLoading && <Loader />}
+        {this.props.products.length === 0 && <Loader />}
         <ProductsWrap>
           {this.props.products.map(product => (
             <Product
