+++ src/store/cart/actions.js
@@ -1,5 +1,5 @@
-export const ADD_PRODUCT = 'cartItems/ADD'
-export const REMOVE_PRODUCT = 'cartItems/REMOVE'
+export const ADD_PRODUCT = 'cart/ADD_PRODUCT'
+export const REMOVE_PRODUCT = 'cart/REMOVE_PRODUCT'
 
 export const addProduct = productId => ({
   type: ADD_PRODUCT,
