+++ src/store/products/actions.js
@@ -1,5 +1,5 @@
-export const LOAD_PRODUCTS = 'products/LOAD'
-export const LOAD_PRODUCT = 'products/LOAD_ONE'
+export const LOAD_PRODUCTS = 'products/LOAD_PRODUCTS'
+export const LOAD_PRODUCT = 'products/LOAD_PRODUCT'
 
 export const loadProducts = products => ({
   type: LOAD_PRODUCTS,
